% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{convert_to_sf}
\alias{convert_to_sf}
\alias{sf_convert}
\title{convert_to_sf}
\usage{
convert_to_sf(x)

sf_convert(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
The converted character vector
}
\description{
Converts a character vector to a stringfish vector
}
\details{
Converts a character vector to a stringfish vector. The opposite of `materialize`.
}
\examples{
x <- convert_to_sf(letters)
}
