% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{convert_to_sf_vector}
\alias{convert_to_sf_vector}
\title{convert_to_sf_vector}
\usage{
convert_to_sf_vector(x, length.out = length(x))
}
\arguments{
\item{x}{A character vector}

\item{length.out}{Optional output length used to recycle `x`}
}
\value{
The converted character vector
}
\description{
Converts a character vector to an `sf_vec`-backed stringfish vector
}
\details{
Converts a character vector to a stringfish vector backed by `sf_vec`.
If `length.out` is supplied, `x` is recycled to that length before conversion.
The opposite of `materialize`.
}
\examples{
x <- convert_to_sf_vector(letters)
}
