% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\name{get_string_type}
\alias{get_string_type}
\title{get_string_type}
\usage{
get_string_type(x)
}
\arguments{
\item{x}{the vector}
}
\value{
The type of vector
}
\description{
Returns the type of the character vector
}
\details{
A function that returns the type of character vector. Possible values are "normal vector", "stringfish vector", "stringfish vector (materialized)" or "other alt-rep vector"
}
\examples{
x <- sf_vector(10)
get_string_type(x) # returns "stringfish vector"
x <- character(10)
get_string_type(x) # returns "normal vector"
}
