% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\name{inspect}
\alias{inspect}
\title{inspect}
\usage{
inspect(x)
}
\arguments{
\item{x}{An R object}
}
\value{
Returns x invisibly
}
\description{
A simple utility function that calls R's internal inspect function
}
\details{
The function prints out the call to the internal inspect function. I.e. `.Internal(inspect(x))`.
}
\examples{
x <- sf_vector(10)
inspect(x)
}
