% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\name{materialize}
\alias{materialize}
\title{materialize}
\usage{
materialize(x)
}
\arguments{
\item{x}{An alt-rep object}
}
\value{
x
}
\description{
Materializes an alt-rep object
}
\details{
Materializes any alt-rep object and then returns it. 
Note: the object is materialized regardless of whether the return value is assigned to a variable.
}
\examples{
x <- sf_vector(10)
sf_assign(x, 1, "hello world")
sf_assign(x, 2, "another string")
x <- materialize(x)
}
