% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{random_strings}
\alias{random_strings}
\title{random_strings}
\usage{
random_strings(N, string_size = 50L, charset = "abcdefghijklmnopqrstuvwxyz",
                      vector_mode = "stringfish")
}
\arguments{
\item{N}{The number of strings to generate}

\item{string_size}{Either a single non-negative integer applied to every output
string, or a non-negative integer vector of length `N`.}

\item{charset}{The characters used to generate the random strings (default: abcdefghijklmnopqrstuvwxyz)}

\item{vector_mode}{The type of character vector to generate (either stringfish or normal, default: stringfish)}
}
\value{
A character vector of the random strings
}
\description{
A function that generates random strings
}
\details{
A convenience function for generating test strings.
}
\examples{
set.seed(1)
x <- random_strings(1e6, 80L, "ACGT", vector_mode = "stringfish")
y <- random_strings(4, c(1L, 2L, 4L, 8L), "ACGT")
}
