% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\name{sf_assign}
\alias{sf_assign}
\title{sf_assign}
\usage{
sf_assign(x, i, e)
}
\arguments{
\item{x}{the vector}

\item{i}{the index to assign to}

\item{e}{the new string to replace at i in x}
}
\description{
Assigns a new string to a stringfish vector or any other character vector
}
\details{
A function to assign a new element to an existing character vector. If the the vector is a stringfish vector, it does so without materialization.
}
\examples{
x <- sf_vector(10)
sf_assign(x, i, "hello world")
}
