% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{sf_ends}
\alias{sf_ends}
\title{sf_ends}
\usage{
sf_ends(subject, pattern, ...)
}
\arguments{
\item{subject}{A character vector}

\item{pattern}{A string to look for at the start}

\item{...}{Parameters passed to sf_grepl}
}
\value{
A logical vector true if there is a match, false if no match, NA is the subject was NA
}
\description{
A function for detecting a pattern at the end of a string
}
\examples{
x <- c("alpha", "beta", "gamma", "delta", "epsilon")
sf_ends(x, "a")
}
\seealso{
endsWith, sf_starts
}
