% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\name{sf_iconv}
\alias{sf_iconv}
\title{sf_iconv}
\usage{
sf_iconv(x, from, to)
}
\arguments{
\item{x}{An alt-rep object}

\item{from}{the encoding to assume of `x`}

\item{to}{the new encoding}
}
\value{
the converted character vector as a stringfish vector
}
\description{
Converts encoding of one character vector to another
}
\details{
This is an analogue to the base R function `iconv`. It converts a string from one encoding (e.g. latin1 or UTF-8) to another
}
\examples{
x <- "fa\xE7ile"
Encoding(x) <- "latin1"
sf_iconv(x, "latin1", "UTF-8")
}
\seealso{
iconv
}
