% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{sf_match}
\alias{sf_match}
\title{sf_match}
\usage{
sf_match(x, table, nthreads = 1)
}
\arguments{
\item{x}{A character vector to search for in table}

\item{table}{A character vector to be matched against x}

\item{nthreads}{Number of threads to use}
}
\value{
An integer vector of the indicies of each x element's position in table
}
\description{
Returns a vector of the positions of x in table
}
\details{
Note: similarly to the base R function, long "table" vectors are not supported. This is due to the maximum integer value that can be returned (`.Machine$integer.max`)
}
\examples{
sf_match("c", letters)
}
\seealso{
match
}
