% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\name{sf_paste}
\alias{sf_paste}
\title{sf_paste}
\usage{
sf_paste(..., sep = "")
}
\arguments{
\item{...}{Any number of character vector strings}

\item{sep}{The seperating string between strings}
}
\value{
A character vector where elements of the arguments are pasted together
}
\description{
Pastes a series of strings together
}
\details{
This works the same way as `paste0(..., sep=sep)`
}
\examples{
x <- letters
y <- LETTERS
sf_paste(x,y, sep = ":")
}
\seealso{
paste0, paste
}
