% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\name{sf_readLines}
\alias{sf_readLines}
\title{sf_readLines}
\usage{
sf_readLines(file, encoding = "UTF-8")
}
\arguments{
\item{file}{The file name}

\item{encoding}{The encoding to use (Default: UTF-8)}
}
\value{
A stringfish vector of the lines in a file
}
\description{
A function that reads a file line by line
}
\details{
A function for reading in text data using `std::ifstream`.
}
\examples{
file <- tempfile()
writeLines(letters, file)
sf_readLines(file)
}
\seealso{
readLines
}
