% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\name{sf_split}
\alias{sf_split}
\title{sf_split}
\usage{
sf_split(subject, split, encode_mode = "auto", fixed = false)

string_identical(x, y)
}
\arguments{
\item{subject}{A character vector}

\item{split}{A delimiter to split the string by}

\item{encode_mode}{"auto", "UTF-8" or "byte". Determines multi-byte (UTF-8) characters or single-byte characters are used.}

\item{fixed}{determines whether the split parameter should be interpreted literally or as a regular expression}

\item{x}{A character vector}

\item{y}{Another character to compare to x}
}
\value{
A list of stringfish character vectors

TRUE if
}
\description{
A function to split strings by a delimiter

A stricter comparison of string equality
}
\examples{
sf_split(datasets::state.name, "\\\\W") # split U.S. state names by any space character
}
\seealso{
strsplit

identical
}
