% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{sf_starts}
\alias{sf_starts}
\title{sf_starts}
\usage{
sf_starts(subject, pattern, ...)
}
\arguments{
\item{subject}{A character vector}

\item{pattern}{A string to look for at the start}

\item{...}{Parameters passed to sf_grepl}
}
\value{
A logical vector true if there is a match, false if no match, NA is the subject was NA
}
\description{
A function for detecting a pattern at the start of a string
}
\examples{
x <- c("alpha", "beta", "gamma", "delta", "epsilon")
sf_starts(x, "a")
}
\seealso{
startsWith, sf_ends
}
