% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{sf_tolower}
\alias{sf_tolower}
\title{sf_tolower}
\usage{
sf_tolower(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A stringfish vector where all uppercase is converted to lowercase
}
\description{
A function converting a string to all lowercase
}
\details{
Note: the function only converts ASCII characters.
}
\examples{
x <- LETTERS
sf_tolower(x)
}
\seealso{
tolower
}
