% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{sf_toupper}
\alias{sf_toupper}
\title{sf_toupper}
\usage{
sf_toupper(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A stringfish vector where all lowercase is converted to uppercase
}
\description{
A function converting a string to all uppercase
}
\details{
Note: the function only converts ASCII characters.
}
\examples{
x <- letters
sf_toupper(x)
}
\seealso{
toupper
}
