% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\name{sf_trim}
\alias{sf_trim}
\title{sf_trim}
\usage{
sf_starts(subject, which = "both", whitespace = "[ \\t\\r\\n]")
}
\arguments{
\item{subject}{A character vector}

\item{which}{"Both", "left", or "right" determines which white space is removed}

\item{whitespace}{Whitespace characters (default: "[ \\t\\r\\n]")}

\item{...}{Parameters passed to sf_gsub}
}
\value{
A stringfish vector of trimmed whitespace
}
\description{
A function to remove leading/trailing whitespace
}
\examples{
x <- c(" alpha ", " beta", " gamma ", "delta ", "epsilon ")
sf_trim(x)
}
\seealso{
trimws
}
