% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{sf_vector}
\alias{sf_vector}
\title{sf_vector}
\usage{
sf_vector(len)
}
\arguments{
\item{len}{length of the new vector}
}
\value{
A new empty stringfish vector
}
\description{
Creates a new empty stringfish vector
}
\details{
This is a backwards-compatible alias for `sf_vector_create(len)`.
It creates a new stringfish vector, an alt-rep character vector backed by a C++ "std::vector" as the internal memory representation.
The vector type is "sfstring", which is a simple C++ class containing owned string bytes and a single byte (uint8_t) representing the encoding.
}
\examples{
x <- sf_vector(10)
sf_assign(x, 1, "hello world")
sf_assign(x, 2, "another string")
}
