% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{sf_vector_create}
\alias{sf_vector_create}
\title{sf_vector_create}
\usage{
sf_vector_create(len)
}
\arguments{
\item{len}{length of the new vector}
}
\value{
A new stringfish vector
}
\description{
Creates a new empty stringfish vector
}
\details{
This function creates a new empty `sf_vec`-backed stringfish vector.
If you want to fill the vector from character data, use `convert_to_sf_vector`.
}
\examples{
x <- sf_vector_create(4)
}
