% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{sf_writeLines}
\alias{sf_writeLines}
\title{sf_writeLines}
\usage{
sf_writeLines(text, file, sep = "\n", na_value = "NA", encode_mode = "UTF-8")
}
\arguments{
\item{text}{A character to write to file}

\item{file}{Name of the file to write to}

\item{sep}{The line separator character(s)}

\item{na_value}{What to write in case of a NA string}

\item{encode_mode}{"UTF-8" or "byte". If "UTF-8", all strings are re-encoded as UTF-8.}
}
\description{
A function that reads a file line by line
}
\details{
A function for writing text data using `std::ofstream`.
}
\examples{
file <- tempfile()
sf_writeLines(letters, file)
sf_readLines(file)
}
\seealso{
writeLines
}
