% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{slice_store_create}
\alias{slice_store_create}
\title{slice_store_create}
\usage{
slice_store_create(len)
}
\arguments{
\item{len}{length of the new vector}
}
\value{
A new slice-store-backed stringfish vector
}
\description{
Creates a new empty slice-store-backed stringfish vector
}
\details{
This function creates a new stringfish vector backed by `slice_store`,
which stores string bytes in append-only slices plus per-element records.
If you want to fill the vector from character data, use `convert_to_slice_store`.
}
\examples{
x <- slice_store_create(4)
}
