% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{slice_store_create_with_size}
\alias{slice_store_create_with_size}
\title{slice_store_create_with_size}
\usage{
slice_store_create_with_size(len, initial_slice_size)
}
\arguments{
\item{len}{length of the new vector}

\item{initial_slice_size}{Initial size of the first underlying `slice_store` slice.}
}
\value{
A new slice-store-backed stringfish vector
}
\description{
Creates a new empty slice-store-backed stringfish vector with a fixed initial slice size
}
\details{
This function creates a new stringfish vector backed by `slice_store`,
and uses `initial_slice_size` for the first slice allocation instead of the default heuristic.
If you want to fill the vector from character data, use `convert_to_slice_store`.
}
\examples{
x <- slice_store_create_with_size(4, 256)
}
