% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{sf_compare}
\alias{sf_compare}
\alias{sf_equals}
\title{sf_compare}
\usage{
sf_compare(x, y, nthreads = 1)

sf_equals(x, y, nthreads = 1L)
}
\arguments{
\item{x}{A character vector of length 1 or the same non-zero length as y}

\item{y}{Another character vector of length 1 or the same non-zero length as y}

\item{nthreads}{Number of threads to use}
}
\value{
A logical vector
}
\description{
Returns a logical vector testing equality of strings from two string vectors
}
\details{
Note: the function tests for both string and encoding equality
}
\examples{
sf_compare(letters, "a")
}
