% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{sf_split}
\alias{sf_split}
\title{sf_split}
\usage{
sf_split(subject, split, encode_mode = "auto", fixed = FALSE, nthreads = 1)
}
\arguments{
\item{subject}{A character vector}

\item{split}{A delimiter to split the string by}

\item{encode_mode}{"auto", "UTF-8" or "byte". Determines multi-byte (UTF-8) characters or single-byte characters are used.}

\item{fixed}{determines whether the split parameter should be interpreted literally or as a regular expression}

\item{nthreads}{Number of threads to use}
}
\value{
A list of stringfish character vectors
}
\description{
A function to split strings by a delimiter
}
\examples{
sf_split(datasets::state.name, "\\\\s") # split U.S. state names by any space character
}
\seealso{
strsplit
}
