% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_files.R
\name{string_identical}
\alias{string_identical}
\title{string_identical}
\usage{
string_identical(x, y)
}
\arguments{
\item{x}{A character vector}

\item{y}{Another character to compare to x}
}
\value{
TRUE if
}
\description{
A stricter comparison of string equality
}
\seealso{
identical
}
